[refactor] separate components
