+++ apps/extension/src/pages/register/enable-chains/components/enable-chains-arrow-down-icon.tsx
@@ -1,0 +1,29 @@
+import React from "react";
+
+export const EnableChainsArrowDownIcon = ({
+  width = "1rem",
+  height = "1rem",
+  color,
+}: {
+  width?: string;
+  height?: string;
+  color?: string;
+}) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={width}
+      height={height}
+      viewBox="0 0 16 17"
+      fill="none"
+    >
+      <path
+        d="M13 6L8 11L3 6"
+        stroke={color || "currentColor"}
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
