+++ apps/extension/src/pages/register/enable-chains/index.tsx
@@ -18,45 +18,32 @@ import {
 import { ChainInfo, ModularChainInfo } from "@keplr-wallet/types";
 import { CoinPretty, Dec } from "@keplr-wallet/unit";
 import { Box } from "../../../components/box";
-import { Column, Columns } from "../../../components/column";
-import { XAxis, YAxis } from "../../../components/axis";
+import { XAxis } from "../../../components/axis";
 import { Gutter } from "../../../components/gutter";
 import { SearchTextInput } from "../../../components/input";
-import {
-  BaseTypography,
-  Body3,
-  Subtitle2,
-  Subtitle3,
-  Subtitle4,
-} from "../../../components/typography";
+import { Body3, Subtitle3, Subtitle4 } from "../../../components/typography";
 import { Button } from "../../../components/button";
 import { ColorPalette } from "../../../styles";
 import { useEffectOnce } from "../../../hooks/use-effect-once";
 import { useNavigate } from "react-router";
-import {
-  ChainImageFallback,
-  CurrencyImageFallback,
-} from "../../../components/image";
-import { Checkbox } from "../../../components/checkbox";
+import { ChainImageFallback } from "../../../components/image";
 import { KeyRingCosmosService } from "@keplr-wallet/background";
 import { WalletStatus } from "@keplr-wallet/stores";
 import { ChainIdHelper } from "@keplr-wallet/cosmos";
 import { TextButton } from "../../../components/button-text";
 import { FormattedMessage, useIntl } from "react-intl";
-import { Tag } from "../../../components/tag";
 import SimpleBar from "simplebar-react";
 import { useTheme } from "styled-components";
 import { dispatchGlobalEventExceptSelf } from "../../../utils/global-events";
 import { VerticalCollapseTransition } from "../../../components/transition/vertical-collapse";
-import { NativeChainMarkIcon } from "../../../components/icon";
 import { EmbedChainInfos } from "../../../config";
-import { ViewToken } from "../../main";
-import { IconButton } from "../../../components/icon-button";
-import { DenomHelper } from "@keplr-wallet/common";
-import { Tooltip } from "../../../components/tooltip";
-import { useSpring, animated } from "@react-spring/web";
 import { useGetAllNonNativeChain } from "../../../hooks/use-get-all-non-native-chain";
 import { hexToRgba } from "../../../utils";
+import { ChainItem } from "./components/chain-item";
+import { NativeChainSection } from "./components/native-chain-section";
+import { NativeChainSectionIconDM } from "./components/native-chain-section-icon-dm";
+import { NativeChainSectionIconLM } from "./components/native-chain-section-icon-lm";
+import { NextStepChainItem } from "./components/next-step-chain-item";
 
 /**
  * EnableChainsScene은 finalize-key scene에서 선택한 chains를 활성화하는 scene이다.
@@ -1803,711 +1790,3 @@ export const EnableChainsScene: FunctionComponent<{
     );
   }
 );
-
-const NativeChainSection: FunctionComponent<{
-  isCollapsed: boolean;
-  isSelectAll: boolean;
-  children: React.ReactNode;
-
-  showTitleBox?: boolean;
-  enabledNativeChainIdentifierList: string[];
-  onClickCollapse: () => void;
-  onClick: () => void;
-  title: string;
-  titleChainImage: React.ReactNode;
-  paragraph: string;
-}> = ({
-  isCollapsed,
-  isSelectAll,
-  children,
-  enabledNativeChainIdentifierList,
-  onClick,
-  onClickCollapse,
-  showTitleBox,
-  title,
-  titleChainImage,
-  paragraph,
-}) => {
-  const theme = useTheme();
-  const nativeChainMarkColor = isSelectAll
-    ? theme.mode === "light"
-      ? ColorPalette["gray-10"]
-      : ColorPalette["gray-550"]
-    : theme.mode === "light"
-    ? ColorPalette.white
-    : ColorPalette["gray-600"];
-
-  return (
-    <React.Fragment>
-      {showTitleBox && (
-        <Box
-          borderRadius="0.375rem"
-          paddingX="1rem"
-          paddingY="0.75rem"
-          borderWidth="1.5px"
-          borderColor={
-            theme.mode === "light"
-              ? ColorPalette["blue-200"]
-              : ColorPalette["blue-600"]
-          }
-          backgroundColor={
-            isSelectAll
-              ? theme.mode === "light"
-                ? ColorPalette["gray-10"]
-                : ColorPalette["gray-550"]
-              : theme.mode === "light"
-              ? ColorPalette.white
-              : ColorPalette["gray-600"]
-          }
-          cursor={"pointer"}
-          onClick={onClick}
-        >
-          <Columns sum={1} alignY="center">
-            {isCollapsed ? (
-              <NativeChainMarkIcon
-                width="1.75rem"
-                height="1.75rem"
-                color={nativeChainMarkColor}
-              />
-            ) : (
-              <Box position="relative">
-                {titleChainImage}
-
-                <Box
-                  position="absolute"
-                  style={{
-                    bottom: "-0.125rem",
-                    right: "-0.125rem",
-                  }}
-                >
-                  <NativeChainMarkIcon
-                    width="1.25rem"
-                    height="1.25rem"
-                    color={nativeChainMarkColor}
-                  />
-                </Box>
-              </Box>
-            )}
-            <Gutter size="0.5rem" />
-            <YAxis>
-              <XAxis alignY="center">
-                <Subtitle2>{title}</Subtitle2>
-                <Gutter size="0.25rem" />
-                <Subtitle3 color={ColorPalette["blue-300"]}>
-                  {enabledNativeChainIdentifierList.length}
-                </Subtitle3>
-              </XAxis>
-              <Gutter size="0.25rem" />
-              {isCollapsed ? null : (
-                <Subtitle3 color={ColorPalette["gray-300"]}>
-                  {" "}
-                  {paragraph}
-                </Subtitle3>
-              )}
-            </YAxis>
-            <Column weight={1} />
-
-            <Column weight={1} />
-            <XAxis alignY="center">
-              <IconButton
-                onClick={(e) => {
-                  e.stopPropagation();
-                  onClickCollapse();
-                }}
-                padding="0.25rem"
-                hoverColor={
-                  theme.mode === "light"
-                    ? hexToRgba(ColorPalette["gray-100"], 0.5)
-                    : ColorPalette["gray-500"]
-                }
-              >
-                {isCollapsed ? (
-                  <LocalArrowDownIcon
-                    width="1.5rem"
-                    height="1.5rem"
-                    color={
-                      theme.mode === "light"
-                        ? ColorPalette["gray-200"]
-                        : ColorPalette["gray-300"]
-                    }
-                  />
-                ) : (
-                  <LocalArrowUpIcon
-                    width="1.5rem"
-                    height="1.5rem"
-                    color={
-                      theme.mode === "light"
-                        ? ColorPalette["gray-200"]
-                        : ColorPalette["gray-300"]
-                    }
-                  />
-                )}
-              </IconButton>
-              <Gutter size="0.5rem" />
-              <Checkbox
-                checked={isSelectAll}
-                onChange={() => {
-                  onClick();
-                }}
-              />
-            </XAxis>
-          </Columns>
-        </Box>
-      )}
-      <VerticalCollapseTransition
-        collapsed={showTitleBox ? isCollapsed : false}
-        opacityLeft={0}
-        transitionAlign="top"
-      >
-        <Gutter size="0.5rem" />
-        <Stack gutter="0.5rem">{children}</Stack>
-      </VerticalCollapseTransition>
-    </React.Fragment>
-  );
-};
-
-const ChainItem: FunctionComponent<{
-  chainInfo: ChainInfo | ModularChainInfo;
-  balance?: CoinPretty;
-  isNativeChain?: boolean;
-  enabled: boolean;
-  blockInteraction: boolean;
-
-  onClick: () => void;
-
-  isFresh: boolean;
-  tokens?: ViewToken[];
-  showTagText?: "EVM" | "Starknet";
-}> = observer(
-  ({
-    chainInfo,
-    balance,
-    enabled,
-    blockInteraction,
-    isNativeChain,
-    onClick,
-    isFresh,
-    tokens,
-    showTagText,
-  }) => {
-    const { priceStore } = useStore();
-    const theme = useTheme();
-    const [isCollapsedTokenView, setIsCollapsedTokenView] = useState(true);
-
-    const price = balance ? priceStore.calculatePrice(balance) : undefined;
-
-    const chainIdentifier = ChainIdHelper.parse(chainInfo.chainId).identifier;
-
-    const arrowAnimation = useSpring({
-      transform: isCollapsedTokenView ? "rotate(0deg)" : "rotate(-180deg)",
-      config: { tension: 300, friction: 25, clamp: true },
-    });
-
-    const isShowTokenView =
-      !!tokens?.length &&
-      (!balance?.toDec().isZero() ||
-        tokens.some((token) => !token.token.toDec().isZero()));
-
-    return (
-      <Box
-        borderRadius="0.375rem"
-        paddingX="1rem"
-        paddingY="0.75rem"
-        backgroundColor={
-          enabled
-            ? theme.mode === "light"
-              ? ColorPalette["gray-10"]
-              : ColorPalette["gray-550"]
-            : theme.mode === "light"
-            ? ColorPalette.white
-            : ColorPalette["gray-600"]
-        }
-        cursor={blockInteraction ? "not-allowed" : "pointer"}
-        onClick={() => {
-          if (!blockInteraction) {
-            onClick();
-          }
-        }}
-      >
-        <Columns sum={1}>
-          <XAxis alignY="center">
-            {isNativeChain ? (
-              <Box position="relative">
-                <ChainImageFallback chainInfo={chainInfo} size="3rem" />
-                <Box
-                  position="absolute"
-                  style={{
-                    bottom: "-0.125rem",
-                    right: "-0.125rem",
-                  }}
-                >
-                  <NativeChainMarkIcon
-                    width="1.25rem"
-                    height="1.25rem"
-                    color={
-                      enabled
-                        ? theme.mode === "light"
-                          ? ColorPalette["gray-10"]
-                          : ColorPalette["gray-550"]
-                        : theme.mode === "light"
-                        ? ColorPalette.white
-                        : ColorPalette["gray-600"]
-                    }
-                  />
-                </Box>
-              </Box>
-            ) : (
-              <ChainImageFallback chainInfo={chainInfo} size="3rem" />
-            )}
-
-            <Gutter size="0.5rem" />
-
-            <YAxis>
-              <XAxis alignY="center">
-                <XAxis alignY="center">
-                  <Subtitle2>
-                    {(() => {
-                      // Noble의 경우만 약간 특수하게 표시해줌
-                      if (chainIdentifier === "noble") {
-                        return `${chainInfo.chainName} (USDC)`;
-                      }
-
-                      return chainInfo.chainName;
-                    })()}
-                  </Subtitle2>
-
-                  {showTagText && (
-                    <React.Fragment>
-                      <Gutter size="0.375rem" />
-                      <Tag text={showTagText} />
-                    </React.Fragment>
-                  )}
-                </XAxis>
-                {isShowTokenView && (
-                  <React.Fragment>
-                    <Gutter size="0.25rem" />
-                    <IconButton
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        e.preventDefault();
-                        setIsCollapsedTokenView(!isCollapsedTokenView);
-                      }}
-                      hoverColor={
-                        theme.mode === "light"
-                          ? hexToRgba(ColorPalette["gray-100"], 0.5)
-                          : ColorPalette["gray-500"]
-                      }
-                      padding="0.25rem"
-                    >
-                      <animated.div
-                        style={{
-                          ...arrowAnimation,
-                          padding: "0",
-                          height: "1rem",
-                        }}
-                      >
-                        <LocalArrowDownIcon
-                          width="1rem"
-                          height="1rem"
-                          color={
-                            theme.mode === "light"
-                              ? ColorPalette["gray-200"]
-                              : ColorPalette["gray-300"]
-                          }
-                        />
-                      </animated.div>
-                    </IconButton>
-                  </React.Fragment>
-                )}
-              </XAxis>
-              {isShowTokenView && (
-                <React.Fragment>
-                  <Gutter size="0.25rem" />
-                  <Subtitle3 color={ColorPalette["gray-300"]}>
-                    {tokens?.length} Tokens
-                  </Subtitle3>
-                </React.Fragment>
-              )}
-            </YAxis>
-          </XAxis>
-          <Column weight={1} />
-          <XAxis alignY="center">
-            {isFresh ||
-            balance == null ||
-            (balance.toDec().isZero() && !isShowTokenView) ? null : (
-              <YAxis alignX="right">
-                <Subtitle3
-                  color={
-                    theme.mode === "light"
-                      ? ColorPalette["gray-600"]
-                      : ColorPalette.white
-                  }
-                >
-                  {balance
-                    .maxDecimals(6)
-                    .shrink(true)
-                    .inequalitySymbol(true)
-                    .toString()}
-                </Subtitle3>
-                <Gutter size="0.25rem" />
-                <Subtitle3 color={ColorPalette["gray-300"]}>
-                  {price ? price.toString() : "-"}
-                </Subtitle3>
-              </YAxis>
-            )}
-
-            <Gutter size="1rem" />
-            <Checkbox
-              checked={enabled}
-              onChange={() => {
-                if (!blockInteraction) {
-                  onClick();
-                }
-              }}
-            />
-          </XAxis>
-        </Columns>
-        {tokens && (
-          <VerticalCollapseTransition
-            collapsed={isCollapsedTokenView}
-            opacityLeft={0}
-          >
-            <TokenView tokens={tokens} />
-          </VerticalCollapseTransition>
-        )}
-      </Box>
-    );
-  }
-);
-
-const TokenView: FunctionComponent<{
-  tokens: ViewToken[];
-}> = ({ tokens }) => {
-  const theme = useTheme();
-
-  return (
-    <Box
-      paddingX="1rem"
-      paddingY="0.75rem"
-      marginTop="0.625rem"
-      borderRadius="0.375rem"
-      backgroundColor={
-        theme.mode === "light"
-          ? hexToRgba(ColorPalette["gray-100"], 0.5)
-          : ColorPalette["gray-500"]
-      }
-    >
-      <Stack gutter="0.5rem">
-        {tokens?.map((token) => (
-          <TokenItem
-            key={token.token.currency.coinMinimalDenom}
-            viewToken={token}
-          />
-        ))}
-      </Stack>
-    </Box>
-  );
-};
-
-const TokenItem: FunctionComponent<{
-  viewToken: ViewToken;
-}> = ({ viewToken }) => {
-  const theme = useTheme();
-
-  const tag = useMemo(() => {
-    const currency = viewToken.token.currency;
-    const denomHelper = new DenomHelper(currency.coinMinimalDenom);
-    if (
-      denomHelper.type === "native" &&
-      currency.coinMinimalDenom.startsWith("ibc/")
-    ) {
-      return {
-        text: "IBC",
-        tooltip: (() => {
-          const start = currency.coinDenom.indexOf("(");
-          const end = currency.coinDenom.lastIndexOf(")");
-
-          if (start < 0 || end < 0) {
-            return "Unknown";
-          }
-
-          return currency.coinDenom.slice(start + 1, end);
-        })(),
-      };
-    }
-    if (denomHelper.type !== "native") {
-      return {
-        text: denomHelper.type,
-      };
-    }
-  }, [viewToken.token.currency]);
-  const coinDenom = useMemo(() => {
-    if (
-      "originCurrency" in viewToken.token.currency &&
-      viewToken.token.currency.originCurrency
-    ) {
-      return viewToken.token.currency.originCurrency.coinDenom;
-    }
-    return viewToken.token.currency.coinDenom;
-  }, [viewToken.token.currency]);
-
-  return (
-    <Columns
-      alignY="center"
-      sum={1}
-      key={viewToken.token.currency.coinMinimalDenom}
-    >
-      <CurrencyImageFallback
-        chainInfo={viewToken.chainInfo}
-        currency={viewToken.token.currency}
-        size="1.5rem"
-      />
-      <Gutter size="0.5rem" />
-      <Subtitle3
-        color={
-          theme.mode === "light"
-            ? ColorPalette["gray-700"]
-            : ColorPalette["gray-50"]
-        }
-        style={{
-          wordBreak: "break-all",
-        }}
-      >
-        {coinDenom}
-      </Subtitle3>
-
-      {tag ? (
-        <React.Fragment>
-          <Gutter size="0.25rem" />
-          <Box alignY="center" height="1px">
-            <TokenTag text={tag.text} tooltip={tag.tooltip} />
-          </Box>
-        </React.Fragment>
-      ) : null}
-
-      <Column weight={1} />
-      <Gutter size="0.5rem" />
-      <Subtitle3
-        color={
-          theme.mode === "light"
-            ? ColorPalette["gray-300"]
-            : ColorPalette["gray-200"]
-        }
-      >
-        {viewToken.token
-          .hideDenom(true)
-          .maxDecimals(6)
-          .inequalitySymbol(true)
-          .shrink(true)
-          .toString()}
-      </Subtitle3>
-    </Columns>
-  );
-};
-
-const TokenTag: FunctionComponent<{
-  text: string;
-  tooltip?: string;
-}> = ({ text, tooltip }) => {
-  const theme = useTheme();
-
-  return (
-    <Tooltip enabled={!!tooltip} content={tooltip}>
-      <Box
-        alignX="center"
-        alignY="center"
-        backgroundColor={
-          theme.mode === "light"
-            ? ColorPalette["blue-50"]
-            : ColorPalette["gray-550"]
-        }
-        borderRadius="0.375rem"
-        height="1rem"
-        paddingX="0.375rem"
-      >
-        <BaseTypography
-          style={{
-            fontWeight: 400,
-            fontSize: "0.6875rem",
-          }}
-          color={
-            theme.mode === "light"
-              ? ColorPalette["blue-400"]
-              : ColorPalette["gray-200"]
-          }
-        >
-          {text}
-        </BaseTypography>
-      </Box>
-    </Tooltip>
-  );
-};
-
-const NextStepChainItem: FunctionComponent<{
-  modularChainInfo: ModularChainInfo | ChainInfo;
-  tagText: string;
-}> = ({ modularChainInfo, tagText }) => {
-  return (
-    <Box
-      paddingX="1rem"
-      paddingY="0.75rem"
-      cursor="not-allowed"
-      style={{ opacity: 0.5 }}
-    >
-      <Columns sum={1}>
-        <XAxis alignY="center">
-          <ChainImageFallback chainInfo={modularChainInfo} size="3rem" />
-
-          <Gutter size="0.5rem" />
-
-          <YAxis>
-            <XAxis alignY="center">
-              <Subtitle2>{modularChainInfo.chainName}</Subtitle2>
-
-              <Gutter size="0.375rem" />
-
-              <Tag text={tagText} />
-            </XAxis>
-
-            <Gutter size="0.25rem" />
-
-            <Subtitle4 color={ColorPalette["gray-300"]}>
-              {"starknet" in modularChainInfo ? (
-                <FormattedMessage id="pages.register.enable-chains.guide.can-select-starknet-later-step" />
-              ) : (
-                <FormattedMessage id="pages.register.enable-chains.guide.can-select-evm-next-step" />
-              )}
-            </Subtitle4>
-          </YAxis>
-        </XAxis>
-      </Columns>
-    </Box>
-  );
-};
-
-const LocalArrowDownIcon = ({
-  width = "1rem",
-  height = "1rem",
-  color,
-}: {
-  width?: string;
-  height?: string;
-  color?: string;
-}) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width={width}
-      height={height}
-      viewBox="0 0 16 17"
-      fill="none"
-    >
-      <path
-        d="M13 6L8 11L3 6"
-        stroke={color || "currentColor"}
-        strokeWidth="1.5"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-};
-const LocalArrowUpIcon = ({
-  width = "1rem",
-  height = "1rem",
-  color,
-  strokeWidth = "2",
-}: {
-  width?: string;
-  height?: string;
-  color?: string;
-  strokeWidth?: string;
-}) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width={width}
-      height={height}
-      viewBox="0 0 24 25"
-      fill="none"
-    >
-      <path
-        d="M4.5 16.25L12 8.75L19.5 16.25"
-        stroke={color || "currentColor"}
-        strokeWidth={strokeWidth}
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-};
-
-const NativeChainSectionIconLM = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="28"
-      height="28"
-      viewBox="0 0 28 28"
-      fill="none"
-    >
-      <g clipPath="url(#clip0_16278_106923)">
-        <path
-          d="M14 17.5C13.3078 17.5 12.6311 17.2947 12.0555 16.9101C11.4799 16.5256 11.0313 15.9789 10.7664 15.3394C10.5015 14.6999 10.4322 13.9961 10.5673 13.3172C10.7023 12.6383 11.0356 12.0146 11.5251 11.5251C12.0146 11.0356 12.6383 10.7023 13.3172 10.5673C13.9961 10.4322 14.6999 10.5015 15.3394 10.7664C15.9789 11.0313 16.5256 11.4799 16.9101 12.0555C17.2947 12.6311 17.5 13.3078 17.5 14C17.4988 14.9279 17.1297 15.8175 16.4736 16.4736C15.8175 17.1297 14.9279 17.4988 14 17.5ZM14 12.25C13.6539 12.25 13.3155 12.3526 13.0278 12.5449C12.74 12.7372 12.5157 13.0105 12.3832 13.3303C12.2508 13.6501 12.2161 14.0019 12.2836 14.3414C12.3512 14.6809 12.5178 14.9927 12.7626 15.2374C13.0073 15.4822 13.3191 15.6489 13.6586 15.7164C13.9981 15.7839 14.3499 15.7492 14.6697 15.6168C14.9895 15.4843 15.2628 15.26 15.4551 14.9722C15.6474 14.6845 15.75 14.3461 15.75 14C15.75 13.5359 15.5656 13.0908 15.2374 12.7626C14.9092 12.4344 14.4641 12.25 14 12.25ZM4.375 17.5C3.68277 17.5 3.00608 17.2947 2.43051 16.9101C1.85493 16.5256 1.40633 15.9789 1.14142 15.3394C0.876516 14.6999 0.807205 13.9961 0.942253 13.3172C1.0773 12.6383 1.41064 12.0146 1.90013 11.5251C2.38961 11.0356 3.01325 10.7023 3.69219 10.5673C4.37112 10.4322 5.07485 10.5015 5.71439 10.7664C6.35393 11.0313 6.90056 11.4799 7.28514 12.0555C7.66973 12.6311 7.875 13.3078 7.875 14C7.87384 14.9279 7.50472 15.8175 6.8486 16.4736C6.19247 17.1297 5.3029 17.4988 4.375 17.5ZM4.375 12.25C4.02888 12.25 3.69054 12.3526 3.40275 12.5449C3.11497 12.7372 2.89067 13.0105 2.75821 13.3303C2.62576 13.6501 2.5911 14.0019 2.65863 14.3414C2.72615 14.6809 2.89282 14.9927 3.13756 15.2374C3.38231 15.4822 3.69413 15.6489 4.03359 15.7164C4.37306 15.7839 4.72493 15.7492 5.0447 15.6168C5.36447 15.4843 5.63778 15.26 5.83007 14.9722C6.02237 14.6845 6.125 14.3461 6.125 14C6.125 13.5359 5.94063 13.0908 5.61244 12.7626C5.28425 12.4344 4.83913 12.25 4.375 12.25ZM8.75 27.125C8.05777 27.125 7.38108 26.9197 6.80551 26.5351C6.22993 26.1506 5.78133 25.6039 5.51642 24.9644C5.25152 24.3249 5.18221 23.6211 5.31725 22.9422C5.4523 22.2633 5.78564 21.6396 6.27513 21.1501C6.76461 20.6606 7.38825 20.3273 8.06719 20.1923C8.74612 20.0572 9.44985 20.1265 10.0894 20.3914C10.7289 20.6563 11.2756 21.1049 11.6601 21.6805C12.0447 22.2561 12.25 22.9328 12.25 23.625C12.2488 24.5529 11.8797 25.4425 11.2236 26.0986C10.5675 26.7547 9.6779 27.1238 8.75 27.125ZM8.75 21.875C8.40388 21.875 8.06554 21.9776 7.77775 22.1699C7.48997 22.3622 7.26567 22.6355 7.13321 22.9553C7.00076 23.2751 6.9661 23.6269 7.03363 23.9664C7.10115 24.3059 7.26782 24.6177 7.51257 24.8624C7.75731 25.1072 8.06913 25.2739 8.40859 25.3414C8.74806 25.4089 9.09993 25.3742 9.4197 25.2418C9.73947 25.1093 10.0128 24.885 10.2051 24.5972C10.3974 24.3095 10.5 23.9711 10.5 23.625C10.5 23.1609 10.3156 22.7158 9.98744 22.3876C9.65925 22.0594 9.21413 21.875 8.75 21.875ZM19.25 27.125C18.5578 27.125 17.8811 26.9197 17.3055 26.5351C16.7299 26.1506 16.2813 25.6039 16.0164 24.9644C15.7515 24.3249 15.6822 23.6211 15.8173 22.9422C15.9523 22.2633 16.2856 21.6396 16.7751 21.1501C17.2646 20.6606 17.8883 20.3273 18.5672 20.1923C19.2461 20.0572 19.9499 20.1265 20.5894 20.3914C21.2289 20.6563 21.7756 21.1049 22.1601 21.6805C22.5447 22.2561 22.75 22.9328 22.75 23.625C22.7488 24.5529 22.3797 25.4425 21.7236 26.0986C21.0675 26.7547 20.1779 27.1238 19.25 27.125ZM19.25 21.875C18.9039 21.875 18.5655 21.9776 18.2778 22.1699C17.99 22.3622 17.7657 22.6355 17.6332 22.9553C17.5008 23.2751 17.4661 23.6269 17.5336 23.9664C17.6012 24.3059 17.7678 24.6177 18.0126 24.8624C18.2573 25.1072 18.5691 25.2739 18.9086 25.3414C19.2481 25.4089 19.5999 25.3742 19.9197 25.2418C20.2395 25.1093 20.5128 24.885 20.7051 24.5972C20.8974 24.3095 21 23.9711 21 23.625C21 23.1609 20.8156 22.7158 20.4874 22.3876C20.1593 22.0594 19.7141 21.875 19.25 21.875ZM23.625 17.5C22.9328 17.5 22.2561 17.2947 21.6805 16.9101C21.1049 16.5256 20.6563 15.9789 20.3914 15.3394C20.1265 14.6999 20.0572 13.9961 20.1923 13.3172C20.3273 12.6383 20.6606 12.0146 21.1501 11.5251C21.6396 11.0356 22.2633 10.7023 22.9422 10.5673C23.6211 10.4322 24.3249 10.5015 24.9644 10.7664C25.6039 11.0313 26.1506 11.4799 26.5351 12.0555C26.9197 12.6311 27.125 13.3078 27.125 14C27.1238 14.9279 26.7547 15.8175 26.0986 16.4736C25.4425 17.1297 24.5529 17.4988 23.625 17.5ZM23.625 12.25C23.2789 12.25 22.9405 12.3526 22.6528 12.5449C22.365 12.7372 22.1407 13.0105 22.0082 13.3303C21.8758 13.6501 21.8411 14.0019 21.9086 14.3414C21.9762 14.6809 22.1428 14.9927 22.3876 15.2374C22.6323 15.4822 22.9441 15.6489 23.2836 15.7164C23.6231 15.7839 23.9749 15.7492 24.2947 15.6168C24.6145 15.4843 24.8878 15.26 25.0801 14.9722C25.2724 14.6845 25.375 14.3461 25.375 14C25.375 13.5359 25.1906 13.0908 24.8624 12.7626C24.5343 12.4344 24.0891 12.25 23.625 12.25ZM19.25 7.875C18.5578 7.875 17.8811 7.66973 17.3055 7.28514C16.7299 6.90056 16.2813 6.35393 16.0164 5.71439C15.7515 5.07485 15.6822 4.37112 15.8173 3.69219C15.9523 3.01325 16.2856 2.38961 16.7751 1.90013C17.2646 1.41064 17.8883 1.0773 18.5672 0.942253C19.2461 0.807205 19.9499 0.876516 20.5894 1.14142C21.2289 1.40633 21.7756 1.85493 22.1601 2.43051C22.5447 3.00608 22.75 3.68277 22.75 4.375C22.7488 5.3029 22.3797 6.19247 21.7236 6.8486C21.0675 7.50472 20.1779 7.87384 19.25 7.875ZM19.25 2.625C18.9039 2.625 18.5655 2.72764 18.2778 2.91993C17.99 3.11222 17.7657 3.38553 17.6332 3.70531C17.5008 4.02508 17.4661 4.37694 17.5336 4.71641C17.6012 5.05588 17.7678 5.3677 18.0126 5.61244C18.2573 5.85718 18.5691 6.02385 18.9086 6.09138C19.2481 6.1589 19.5999 6.12424 19.9197 5.99179C20.2395 5.85934 20.5128 5.63504 20.7051 5.34725C20.8974 5.05946 21 4.72112 21 4.375C21 3.91087 20.8156 3.46575 20.4874 3.13756C20.1593 2.80938 19.7141 2.625 19.25 2.625ZM8.75 7.875C8.05777 7.875 7.38108 7.66973 6.80551 7.28514C6.22993 6.90056 5.78133 6.35393 5.51642 5.71439C5.25152 5.07485 5.18221 4.37112 5.31725 3.69219C5.4523 3.01325 5.78564 2.38961 6.27513 1.90013C6.76461 1.41064 7.38825 1.0773 8.06719 0.942253C8.74612 0.807205 9.44985 0.876516 10.0894 1.14142C10.7289 1.40633 11.2756 1.85493 11.6601 2.43051C12.0447 3.00608 12.25 3.68277 12.25 4.375C12.2488 5.3029 11.8797 6.19247 11.2236 6.8486C10.5675 7.50472 9.6779 7.87384 8.75 7.875ZM8.75 2.625C8.40388 2.625 8.06554 2.72764 7.77775 2.91993C7.48997 3.11222 7.26567 3.38553 7.13321 3.70531C7.00076 4.02508 6.9661 4.37694 7.03363 4.71641C7.10115 5.05588 7.26782 5.3677 7.51257 5.61244C7.75731 5.85718 8.06913 6.02385 8.40859 6.09138C8.74806 6.1589 9.09993 6.12424 9.4197 5.99179C9.73947 5.85934 10.0128 5.63504 10.2051 5.34725C10.3974 5.05946 10.5 4.72112 10.5 4.375C10.5 3.91087 10.3156 3.46575 9.98744 3.13756C9.65925 2.80938 9.21413 2.625 8.75 2.625Z"
-          fill="url(#paint0_linear_16278_106923)"
-        />
-      </g>
-      <defs>
-        <linearGradient
-          id="paint0_linear_16278_106923"
-          x1="14"
-          y1="0.875"
-          x2="14"
-          y2="27.125"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="#9E2EFF" />
-          <stop offset="1" stopColor="#2759F2" />
-        </linearGradient>
-        <clipPath id="clip0_16278_106923">
-          <rect width="28" height="28" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
-
-const NativeChainSectionIconDM = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="28"
-      height="28"
-      viewBox="0 0 28 28"
-      fill="none"
-    >
-      <g clipPath="url(#clip0_16221_98764)">
-        <path
-          d="M14 17.5C13.3078 17.5 12.6311 17.2947 12.0555 16.9101C11.4799 16.5256 11.0313 15.9789 10.7664 15.3394C10.5015 14.6999 10.4322 13.9961 10.5673 13.3172C10.7023 12.6383 11.0356 12.0146 11.5251 11.5251C12.0146 11.0356 12.6383 10.7023 13.3172 10.5673C13.9961 10.4322 14.6999 10.5015 15.3394 10.7664C15.9789 11.0313 16.5256 11.4799 16.9101 12.0555C17.2947 12.6311 17.5 13.3078 17.5 14C17.4988 14.9279 17.1297 15.8175 16.4736 16.4736C15.8175 17.1297 14.9279 17.4988 14 17.5ZM14 12.25C13.6539 12.25 13.3155 12.3526 13.0278 12.5449C12.74 12.7372 12.5157 13.0105 12.3832 13.3303C12.2508 13.6501 12.2161 14.0019 12.2836 14.3414C12.3512 14.6809 12.5178 14.9927 12.7626 15.2374C13.0073 15.4822 13.3191 15.6489 13.6586 15.7164C13.9981 15.7839 14.3499 15.7492 14.6697 15.6168C14.9895 15.4843 15.2628 15.26 15.4551 14.9722C15.6474 14.6845 15.75 14.3461 15.75 14C15.75 13.5359 15.5656 13.0908 15.2374 12.7626C14.9092 12.4344 14.4641 12.25 14 12.25ZM4.375 17.5C3.68277 17.5 3.00608 17.2947 2.43051 16.9101C1.85493 16.5256 1.40633 15.9789 1.14142 15.3394C0.876516 14.6999 0.807205 13.9961 0.942253 13.3172C1.0773 12.6383 1.41064 12.0146 1.90013 11.5251C2.38961 11.0356 3.01325 10.7023 3.69219 10.5673C4.37112 10.4322 5.07485 10.5015 5.71439 10.7664C6.35393 11.0313 6.90056 11.4799 7.28514 12.0555C7.66973 12.6311 7.875 13.3078 7.875 14C7.87384 14.9279 7.50472 15.8175 6.8486 16.4736C6.19247 17.1297 5.3029 17.4988 4.375 17.5ZM4.375 12.25C4.02888 12.25 3.69054 12.3526 3.40275 12.5449C3.11497 12.7372 2.89067 13.0105 2.75821 13.3303C2.62576 13.6501 2.5911 14.0019 2.65863 14.3414C2.72615 14.6809 2.89282 14.9927 3.13756 15.2374C3.38231 15.4822 3.69413 15.6489 4.03359 15.7164C4.37306 15.7839 4.72493 15.7492 5.0447 15.6168C5.36447 15.4843 5.63778 15.26 5.83007 14.9722C6.02237 14.6845 6.125 14.3461 6.125 14C6.125 13.5359 5.94063 13.0908 5.61244 12.7626C5.28425 12.4344 4.83913 12.25 4.375 12.25ZM8.75 27.125C8.05777 27.125 7.38108 26.9197 6.80551 26.5351C6.22993 26.1506 5.78133 25.6039 5.51642 24.9644C5.25152 24.3249 5.18221 23.6211 5.31725 22.9422C5.4523 22.2633 5.78564 21.6396 6.27513 21.1501C6.76461 20.6606 7.38825 20.3273 8.06719 20.1923C8.74612 20.0572 9.44985 20.1265 10.0894 20.3914C10.7289 20.6563 11.2756 21.1049 11.6601 21.6805C12.0447 22.2561 12.25 22.9328 12.25 23.625C12.2488 24.5529 11.8797 25.4425 11.2236 26.0986C10.5675 26.7547 9.6779 27.1238 8.75 27.125ZM8.75 21.875C8.40388 21.875 8.06554 21.9776 7.77775 22.1699C7.48997 22.3622 7.26567 22.6355 7.13321 22.9553C7.00076 23.2751 6.9661 23.6269 7.03363 23.9664C7.10115 24.3059 7.26782 24.6177 7.51257 24.8624C7.75731 25.1072 8.06913 25.2739 8.40859 25.3414C8.74806 25.4089 9.09993 25.3742 9.4197 25.2418C9.73947 25.1093 10.0128 24.885 10.2051 24.5972C10.3974 24.3095 10.5 23.9711 10.5 23.625C10.5 23.1609 10.3156 22.7158 9.98744 22.3876C9.65925 22.0594 9.21413 21.875 8.75 21.875ZM19.25 27.125C18.5578 27.125 17.8811 26.9197 17.3055 26.5351C16.7299 26.1506 16.2813 25.6039 16.0164 24.9644C15.7515 24.3249 15.6822 23.6211 15.8173 22.9422C15.9523 22.2633 16.2856 21.6396 16.7751 21.1501C17.2646 20.6606 17.8883 20.3273 18.5672 20.1923C19.2461 20.0572 19.9499 20.1265 20.5894 20.3914C21.2289 20.6563 21.7756 21.1049 22.1601 21.6805C22.5447 22.2561 22.75 22.9328 22.75 23.625C22.7488 24.5529 22.3797 25.4425 21.7236 26.0986C21.0675 26.7547 20.1779 27.1238 19.25 27.125ZM19.25 21.875C18.9039 21.875 18.5655 21.9776 18.2778 22.1699C17.99 22.3622 17.7657 22.6355 17.6332 22.9553C17.5008 23.2751 17.4661 23.6269 17.5336 23.9664C17.6012 24.3059 17.7678 24.6177 18.0126 24.8624C18.2573 25.1072 18.5691 25.2739 18.9086 25.3414C19.2481 25.4089 19.5999 25.3742 19.9197 25.2418C20.2395 25.1093 20.5128 24.885 20.7051 24.5972C20.8974 24.3095 21 23.9711 21 23.625C21 23.1609 20.8156 22.7158 20.4874 22.3876C20.1593 22.0594 19.7141 21.875 19.25 21.875ZM23.625 17.5C22.9328 17.5 22.2561 17.2947 21.6805 16.9101C21.1049 16.5256 20.6563 15.9789 20.3914 15.3394C20.1265 14.6999 20.0572 13.9961 20.1923 13.3172C20.3273 12.6383 20.6606 12.0146 21.1501 11.5251C21.6396 11.0356 22.2633 10.7023 22.9422 10.5673C23.6211 10.4322 24.3249 10.5015 24.9644 10.7664C25.6039 11.0313 26.1506 11.4799 26.5351 12.0555C26.9197 12.6311 27.125 13.3078 27.125 14C27.1238 14.9279 26.7547 15.8175 26.0986 16.4736C25.4425 17.1297 24.5529 17.4988 23.625 17.5ZM23.625 12.25C23.2789 12.25 22.9405 12.3526 22.6528 12.5449C22.365 12.7372 22.1407 13.0105 22.0082 13.3303C21.8758 13.6501 21.8411 14.0019 21.9086 14.3414C21.9762 14.6809 22.1428 14.9927 22.3876 15.2374C22.6323 15.4822 22.9441 15.6489 23.2836 15.7164C23.6231 15.7839 23.9749 15.7492 24.2947 15.6168C24.6145 15.4843 24.8878 15.26 25.0801 14.9722C25.2724 14.6845 25.375 14.3461 25.375 14C25.375 13.5359 25.1906 13.0908 24.8624 12.7626C24.5343 12.4344 24.0891 12.25 23.625 12.25ZM19.25 7.875C18.5578 7.875 17.8811 7.66973 17.3055 7.28514C16.7299 6.90056 16.2813 6.35393 16.0164 5.71439C15.7515 5.07485 15.6822 4.37112 15.8173 3.69219C15.9523 3.01325 16.2856 2.38961 16.7751 1.90013C17.2646 1.41064 17.8883 1.0773 18.5672 0.942253C19.2461 0.807205 19.9499 0.876516 20.5894 1.14142C21.2289 1.40633 21.7756 1.85493 22.1601 2.43051C22.5447 3.00608 22.75 3.68277 22.75 4.375C22.7488 5.3029 22.3797 6.19247 21.7236 6.8486C21.0675 7.50472 20.1779 7.87384 19.25 7.875ZM19.25 2.625C18.9039 2.625 18.5655 2.72764 18.2778 2.91993C17.99 3.11222 17.7657 3.38553 17.6332 3.70531C17.5008 4.02508 17.4661 4.37694 17.5336 4.71641C17.6012 5.05588 17.7678 5.3677 18.0126 5.61244C18.2573 5.85718 18.5691 6.02385 18.9086 6.09138C19.2481 6.1589 19.5999 6.12424 19.9197 5.99179C20.2395 5.85934 20.5128 5.63504 20.7051 5.34725C20.8974 5.05946 21 4.72112 21 4.375C21 3.91087 20.8156 3.46575 20.4874 3.13756C20.1593 2.80938 19.7141 2.625 19.25 2.625ZM8.75 7.875C8.05777 7.875 7.38108 7.66973 6.80551 7.28514C6.22993 6.90056 5.78133 6.35393 5.51642 5.71439C5.25152 5.07485 5.18221 4.37112 5.31725 3.69219C5.4523 3.01325 5.78564 2.38961 6.27513 1.90013C6.76461 1.41064 7.38825 1.0773 8.06719 0.942253C8.74612 0.807205 9.44985 0.876516 10.0894 1.14142C10.7289 1.40633 11.2756 1.85493 11.6601 2.43051C12.0447 3.00608 12.25 3.68277 12.25 4.375C12.2488 5.3029 11.8797 6.19247 11.2236 6.8486C10.5675 7.50472 9.6779 7.87384 8.75 7.875ZM8.75 2.625C8.40388 2.625 8.06554 2.72764 7.77775 2.91993C7.48997 3.11222 7.26567 3.38553 7.13321 3.70531C7.00076 4.02508 6.9661 4.37694 7.03363 4.71641C7.10115 5.05588 7.26782 5.3677 7.51257 5.61244C7.75731 5.85718 8.06913 6.02385 8.40859 6.09138C8.74806 6.1589 9.09993 6.12424 9.4197 5.99179C9.73947 5.85934 10.0128 5.63504 10.2051 5.34725C10.3974 5.05946 10.5 4.72112 10.5 4.375C10.5 3.91087 10.3156 3.46575 9.98744 3.13756C9.65925 2.80938 9.21413 2.625 8.75 2.625Z"
-          fill="url(#paint0_linear_16221_98764)"
-        />
-      </g>
-      <defs>
-        <linearGradient
-          id="paint0_linear_16221_98764"
-          x1="14"
-          y1="0.875"
-          x2="14"
-          y2="27.125"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="#8D7CFF" />
-          <stop offset="1" stopColor="#2759F2" />
-        </linearGradient>
-        <clipPath id="clip0_16221_98764">
-          <rect width="28" height="28" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
